+++ src/bigdict/__init__.py
@@ -1,7 +1,7 @@
 """
 Bigdict
 """
-__version__ = "0.2.9"
+__version__ = "0.3.0"
 
 from ._bigdict import Bigdict